splitup persistent_processing
