+++ src/actinia_core/rest/persistent_mapset_merger.py
@@ -27,7 +27,8 @@ Asynchronous merging of several mapsets into a single one
 import pickle
 from flask import jsonify, make_response
 
-from actinia_core.rest.persistent_processing import PersistentProcessing
+from actinia_core.processing.actinia_processing.ephemeral.persistent_processing \
+     import PersistentProcessing
 from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.core.common.exceptions import AsyncProcessError, \
