+++ src/actinia_core/rest/raster_layer.py
@@ -36,7 +36,8 @@ from actinia_api.swagger2.actinia_core.schemas.raster_layer import \
 from actinia_core.core.common.app import URL_PREFIX
 from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
      import EphemeralProcessing
-from actinia_core.rest.persistent_processing import PersistentProcessing
+from actinia_core.processing.actinia_processing.ephemeral.persistent_processing \
+     import PersistentProcessing
 from actinia_core.rest.base.map_layer_base import MapLayerRegionResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.core.common.exceptions import AsyncProcessError
