+++ src/actinia_core/rest/resource_storage_management.py
@@ -31,8 +31,10 @@ import os
 from flask import jsonify, make_response
 from flask_restful_swagger_2 import swagger
 import pickle
-from .ephemeral_processing import EphemeralProcessing
-from .persistent_processing import PersistentProcessing
+from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
+     import EphemeralProcessing
+from actinia_core.processing.actinia_processing.ephemeral.persistent_processing \
+     import PersistentProcessing
 from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.core.common.process_object import Process
